extend isinstance for parameterized generics
